feat: BSZ-FRUN-01-runtime-session-surfaces-008 - Preserve Factory Session MCP execution coverage under file size limit
